Finished overworld UI, commit before cleaning project

+++ Assets/WorldEvent.cs
@@ -8,5 +8,5 @@ public class WorldEvent : ScriptableObject {
     public GameObject _markerObject;
     public List<QuickItem> _itemList = new List<QuickItem>();
     public List<WeaponItem> _weaponItemList = new List<WeaponItem>();
-    public float _exp;
+    public int _exp;
 }
